+++ PKHeX.Core/PKM/PA8.cs
@@ -348,9 +348,9 @@ public sealed class PA8 : PKM, ISanityChecksum,
 
     // 0x4C-0x4F unused
 
-    public int HeightScalar { get => Data[0x50]; set => Data[0x50] = (byte)value; }
-    public int WeightScalar { get => Data[0x51]; set => Data[0x51] = (byte)value; }
-    public int HeightScalarCopy { get => Data[0x52]; set => Data[0x52] = (byte)value; }
+    public byte HeightScalar { get => Data[0x50]; set => Data[0x50] = value; }
+    public byte WeightScalar { get => Data[0x51]; set => Data[0x51] = value; }
+    public byte HeightScalarCopy { get => Data[0x52]; set => Data[0x52] = value; }
 
     // 0x53 unused
