+++ PKHeX.Core/PKM/Shared/ICaughtData2.cs
@@ -5,7 +5,7 @@
     /// </summary>
     public interface ICaughtData2
     {
-        int CaughtData { get; set; }
+        ushort CaughtData { get; set; }
         int Met_TimeOfDay { get; set; }
         int Met_Level { get; set; }
         int OT_Gender { get; set; }
